pyproject.toml
[tool.poetry]
name = "ebook-generator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = "^0.63.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
